Add a GitHub Actions pipeline to print command outputs
I am preparing these commands to make sure their outputs match what I expect. In the next commit, I will make the script exit if any of the outputs don't match.

I might need to run the commands in their separate pipelines in order to be able to distinguish what is working and what is not, instead of giving up on the first error.

This would be doable in Bash as well, but it is going to look neater separated like this I think.