Split the pipeline steps into individual pipelines
I want these steps to run independently of one another in case of a failure; they should all always run and if one fails, it should not affect the others.

This could be done with a single pipeline which runs the commands sequentially, collects the errors and if any, exits with a non-zero exit code at the end. But there is no advantage to the sequential order, multiple pipelines will run in parallel (it might not due to GitHub parallelism limits, but that's okay).