Fix neglected job names and add a current directory printing pipeline
I think I am running the scripts with the correct prefix, but it is possible I might need to add ./ before it.

I made this extra pipeline to print the working directory and its tree.

I also changed the job name in the other pipelines to reflect the refactoring into multiple files and the general fact that they all just run a shell script.