Add the missing checkout step
By extracting the shell scripts out of the YAML files, I lost access to them, because I was not checking out the contents of the repository.

This should prove that this was the problem in the pwd pipeline, but I am also experimenting with the @latest tag. I might need to change it to @v3 if @latest is not supported.