Set up artifact upload on failure for the https-ssl-certificates pipeline
This pipeline is currently failing as the agent seems to be getting a different response from my Mac. I need to analyze and diff the files, so I will drop the artifacts on failure so I can download them later.

I will later add this to all pipelines, but while I am tweaking it, I will only add it to this known bad pipeline for now.