Add a badge for the GitHub Pages hidden pipeline
This pipeline is created by enabling GitHub Pages and as such has no file name or workflow name and the format of the badge URL is a little different.

I figured it out by going to the Actions tab, opening the GitHub Pages pipeline and using the triple-dot menu to find the Create Status Badge item. This works for any pipeline and for the GitHub Pages hidden pipeline, it generates this correct, different URL.