Reflect certificate change in the https-ssl-certificates workflow pattern
My noodling around in the GitHub Pages settings page investigating why https://www.hubelbauer.net was throwing security errors caused GitHub to provision a new certificate that covered both the apex and the WWW.

This certificate is different so it stopped matching the old one, hence the pattern update in the workflow.

On top of this, I realized I will need more pipelines around this behavior, so I wrote up what needs to be added and replaced the previous two to-dos which have no longer been relevant knowing this new information.