Move links containing expected DNS record contents to the shell scripts
I kept the Pipelines section in the readme for this information, but in the hindsight, it seems better to co-locate this with the actual checks. I have decided to dump the section completely for now. I might turn the badges into links and add extra comments into the pipelines so they are not completely undocumented, but then again, it's not hard to find them nor is it hard to understand them, so I might do neither.