Add new pipelines that check for various redirects
We're checking for HTTP to HTTPS, HTTP WWW to apex and HTTPS WWW to apex. This replaces the previously deleted pipeline. I've decided to stick with the Location response header check, not the HTML response check. And yes, the Location header is lower-case in one of the pipelines. Not sure why :)