Add new pipelines for redirect domain checks
A bunch of domains point to hubelbauer.net and I should set up checks that make sure this stays the case. I updated the tasks in the readme to take a few notes on what to do with my other domains, which will happen away from this repository, but I don't have a better place to note this yet, so it will stay here for now.