Update SSL certificate pattern
Two or so hours ago the SSL certificate seems to have changed on GitHub's side as I've started getting failures on this workflow. I have compared the actual output with the pattern and the diff revealed a difference in the first certificate in the chain. I've updated it as per the diff so now the workflow should start passing again.